+++ modules/Accounting/Salaries.php
@@ -16,18 +16,8 @@ if ( $_REQUEST['values']
 	&& AllowEdit()
 	&& UserStaffID() )
 {
-	if ( isset( $_POST['day_values'], $_POST['month_values'], $_POST['year_values'] ) )
-	{
-		$requested_dates = RequestedDates(
-			$_REQUEST['year_values'],
-			$_REQUEST['month_values'],
-			$_REQUEST['day_values']
-		);
-
-		$_REQUEST['values'] = array_replace_recursive( (array) $_REQUEST['values'], $requested_dates );
-
-		$_POST['values'] = array_replace_recursive( (array) $_POST['values'], $requested_dates );
-	}
+	// Add eventual Dates to $_REQUEST['values'].
+	AddRequestedDates( 'values', 'post' );
 
 	foreach ( (array) $_REQUEST['values'] as $id => $columns)
 	{