+++ modules/Discipline/Referrals.php
@@ -5,18 +5,8 @@ require_once 'ProgramFunctions/TipMessage.fnc.php';
 
 DrawHeader( ProgramTitle() );
 
-if ( isset( $_POST['day_values'], $_POST['month_values'], $_POST['year_values'] ) )
-{
-	$requested_dates = RequestedDates(
-		$_REQUEST['year_values'],
-		$_REQUEST['month_values'],
-		$_REQUEST['day_values']
-	);
-
-	$_REQUEST['values'] = array_replace_recursive( (array) $_REQUEST['values'], $requested_dates );
-
-	$_POST['values'] = array_replace_recursive( (array) $_POST['values'], $requested_dates );
-}
+// Add eventual Dates to $_REQUEST['values'].
+AddRequestedDates( 'values', 'post' );
 
 if ( isset( $_POST['values'] )
 	&& count( $_POST['values'] )