+++ modules/Student_Billing/StudentPayments.php
@@ -10,16 +10,7 @@ if ( ! $_REQUEST['print_statements'] )
 }
 
 // Add eventual Dates to $_REQUEST['values'].
-if ( isset( $_REQUEST['day_values'], $_REQUEST['month_values'], $_REQUEST['year_values'] ) )
-{
-	$requested_dates = RequestedDates(
-		$_REQUEST['year_values'],
-		$_REQUEST['month_values'],
-		$_REQUEST['day_values']
-	);
-
-	$_REQUEST['values'] = array_replace_recursive( (array) $_REQUEST['values'], (array) $requested_dates );
-}
+AddRequestedDates( 'values', 'post' );
 
 if ( $_REQUEST['values']
 	&& $_POST['values']