+++ modules/Students/AssignOtherInfo.php
@@ -8,18 +8,7 @@ if ( $_REQUEST['modfunc'] === 'save'
 	&& AllowEdit() )
 {
 	// Add eventual Dates to $_REQUEST['values'].
-	if ( isset( $_POST['day_values'], $_POST['month_values'], $_POST['year_values'] ) )
-	{
-		$requested_dates = RequestedDates(
-			$_REQUEST['year_values'],
-			$_REQUEST['month_values'],
-			$_REQUEST['day_values']
-		);
-
-		$_REQUEST['values'] = array_replace_recursive( (array) $_REQUEST['values'], $requested_dates );
-
-		$_POST['values'] = array_replace_recursive( (array) $_POST['values'], $requested_dates );
-	}
+	AddRequestedDates( 'values', 'post' );
 
 	if ( count( $_POST['values'] )
 		&& count( $_POST['student'] ) )